#4: remove obsolete ui-5 fields
