+++ magkit-ui/src/main/java/de/ibmix/magkit/ui/dialogs/fields/SpecificMultiDefinition.java
@@ -1,3 +1,5 @@
+package de.ibmix.magkit.ui.dialogs.fields;
+
 /*-
  * #%L
  * IBM iX Magnolia Kit
@@ -7,9 +9,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,16 +19,17 @@
  * limitations under the License.
  * #L%
  */
-/*Styles for the sortable multivaluefield*/
-.aperto-multifield > .v-expand > .v-slot {
-    width: 80% !important;
-}
 
-.aperto-multifield {
-    margin-bottom: 20px;
-}
+/**
+ * Interface for common specific multi field definitions.
+ *
+ * @author frank.sommer
+ * @since 05.03.2021
+ */
+public interface SpecificMultiDefinition {
+    long DEFAULT_MAX = 3;
 
-.aperto-multi-buttons {
-    top: 10px;
-    position: relative;
+    Long getMaxComponents();
+
+    String getParentComponentProperty();
 }
