+++ magkit-ui/src/main/java/de/ibmix/magkit/ui/dialogs/fields/SpecificMultiFormView.java
@@ -1,4 +1,4 @@
-package de.ibmix.magkit.ui.dialogs.m6.fields;
+package de.ibmix.magkit.ui.dialogs.fields;
 
 /*-
  * #%L
