+++ magkit-ui/src/main/java/de/ibmix/magkit/ui/dialogs/fields/SpecificMultiValueFieldDefinition.java
@@ -1,4 +1,4 @@
-package de.ibmix.magkit.ui.dialogs.m6.fields;
+package de.ibmix.magkit.ui.dialogs.fields;
 
 /*-
  * #%L
@@ -21,20 +21,21 @@ package de.ibmix.magkit.ui.dialogs.m6.fields;
  */
 
 import info.magnolia.ui.field.FieldType;
-import info.magnolia.ui.field.JcrMultiFieldDefinition;
+import info.magnolia.ui.field.JcrMultiValueFieldDefinition;
 
 /**
- * Multi field with specific size. Used for multi fields containing composite fields.
+ * Multi field with specific size. Used for simple multi fields stored in a multi value property.
  *
  * @author payam.tabrizi
  * @since 22.02.21
  */
-@FieldType("specificMultiField")
-public class SpecificMultiFieldDefinition extends JcrMultiFieldDefinition implements SpecificMultiDefinition {
+@FieldType("specificMultiValueField")
+public class SpecificMultiValueFieldDefinition extends JcrMultiValueFieldDefinition implements SpecificMultiDefinition {
+
     private Long _maxComponents = DEFAULT_MAX;
     private String _parentCountProperty;
 
-    public SpecificMultiFieldDefinition() {
+    public SpecificMultiValueFieldDefinition() {
         setImplementationClass((Class) SpecificMultiFormView.class);
     }
