--- magkit-ui/src/main/java/de/ibmix/magkit/ui/dialogs/m6/fields/SpecificMultiDefinition.java
@@ -1,35 +1,0 @@
-package de.ibmix.magkit.ui.dialogs.m6.fields;
-
-/*-
- * #%L
- * IBM iX Magnolia Kit
- * %%
- * Copyright (C) 2023 IBM iX
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-/**
- * Interface for common specific multi field definitions.
- *
- * @author frank.sommer
- * @since 05.03.2021
- */
-public interface SpecificMultiDefinition {
-    long DEFAULT_MAX = 3;
-
-    Long getMaxComponents();
-
-    String getParentComponentProperty();
-}
