--- magkit-ui/src/main/resources/com/aperto/magkit/controls/OrderingMultiSelectMultiRepositoryInner.html
@@ -1,30 +1,0 @@
-<!--
-  #%L
-  IBM iX Magnolia Kit
-  %%
-  Copyright (C) 2023 IBM iX
-  %%
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-  
-       http://www.apache.org/licenses/LICENSE-2.0
-  
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  #L%
-  -->
-<#if this.saveAsList || this.saveAsJSON>
-    <#assign name= ""/>
-<#else/>
-    <#assign name= this.name />
-</#if>
-
-<input onchange="${this.name}DynamicTable.persist();" type="text" id="${r"${prefix}"}" name="${name}" value="${r"${obj.value}"}" class="mgnlDialogControlEdit"/>
-
-${this.chooseButton}
-${this.deleteButton}
-${this.repositorySelect}
