--- magkit-ui/src/main/resources/com/aperto/magkit/dialogs/m6/fields/extendedTextField.css
@@ -1,34 +1,0 @@
-/*-
- * #%L
- * IBM iX Magnolia Kit
- * %%
- * Copyright (C) 2023 IBM iX
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-/* com.aperto.magkit.dialogs.m6.fields.ExtendedTextField */
-.aperto-extended-textfield-label {
-    color: #666;
-    white-space: nowrap;
-    line-height: 11px;
-    right: 5px;
-    position: absolute;
-    font-size: 11px;
-    top: 5px;
-}
-
-.v-customcomponent-error .aperto-extended-textfield input.v-textfield {
-    /* unable to get this working with magnolia styles */
-    border-color: red !important;
-}
