+++ magkit-ui/src/main/resources/magkit/i18n/magkit-messages_de.properties
@@ -7,9 +7,9 @@
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
+#
 #      http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,8 +19,6 @@
 ###
 pageLink.label=Link
 pageLink.description=Interner Link zu einer Seite oder externer Link beginnend mit 'http://' oder 'https://'.
-listSelect.emptySelection.label=- Keine Auswahl -
-
 # custom labels #
 folder.text=Dies ist ein Ordner und enth\u00E4lt keinen Inhalt.
