--- magkit-ui/src/main/resources/mgnl-bootstrap/magkit/config.modules.magkit.fieldTypes.extendedLinkField.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sv:node sv:name="extendedLinkField" xmlns:sv="http://www.jcp.org/jcr/sv/1.0">
-  <sv:property sv:name="jcr:primaryType" sv:type="Name">
-    <sv:value>mgnl:contentNode</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:uuid" sv:type="String">
-    <sv:value>5fd67f24-c110-49d9-a336-1ead5f24181b</sv:value>
-  </sv:property>
-  <sv:property sv:name="definitionClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.ExtendedLinkFieldDefinition</sv:value>
-  </sv:property>
-  <sv:property sv:name="factoryClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.ExtendedLinkFieldFactory</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:createdBy" sv:type="String">
-    <sv:value>admin</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:activationStatus" sv:type="Boolean">
-    <sv:value>false</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:created" sv:type="Date">
-    <sv:value>2015-06-03T16:03:08.050+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:createdBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivated" sv:type="Date">
-    <sv:value>2015-06-03T16:03:08.052+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivatedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModified" sv:type="Date">
-    <sv:value>2015-06-03T16:11:06.155+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModifiedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-</sv:node>
