--- magkit-ui/src/main/resources/mgnl-bootstrap/magkit/config.modules.magkit.fieldTypes.extendedTextField.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sv:node sv:name="extendedTextField" xmlns:sv="http://www.jcp.org/jcr/sv/1.0">
-  <sv:property sv:name="jcr:primaryType" sv:type="Name">
-    <sv:value>mgnl:contentNode</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:uuid" sv:type="String">
-    <sv:value>045c51c6-6563-4ea3-ad91-2c18fd499be3</sv:value>
-  </sv:property>
-  <sv:property sv:name="definitionClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.ExtendedTextFieldDefinition</sv:value>
-  </sv:property>
-  <sv:property sv:name="factoryClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.ExtendedTextFieldFactory</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:createdBy" sv:type="String">
-    <sv:value>admin</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModified" sv:type="Date">
-    <sv:value>2015-03-25T10:45:52.440+01:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModifiedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-</sv:node>
