--- magkit-ui/src/main/resources/mgnl-bootstrap/magkit/config.modules.magkit.fieldTypes.listSelect.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sv:node sv:name="listSelect" xmlns:sv="http://www.jcp.org/jcr/sv/1.0">
-  <sv:property sv:name="jcr:primaryType" sv:type="Name">
-    <sv:value>mgnl:contentNode</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:uuid" sv:type="String">
-    <sv:value>9b9daf8c-edb3-4ef5-aeb4-6cd55eeecdf2</sv:value>
-  </sv:property>
-  <sv:property sv:name="definitionClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.ListSelectFieldDefinition</sv:value>
-  </sv:property>
-  <sv:property sv:name="factoryClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.ListSelectGroupFieldFactory</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:createdBy" sv:type="String">
-    <sv:value>admin</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:activationStatus" sv:type="Boolean">
-    <sv:value>false</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:created" sv:type="Date">
-    <sv:value>2014-10-21T14:01:16.267+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:createdBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivated" sv:type="Date">
-    <sv:value>2014-10-21T14:01:16.268+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivatedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModified" sv:type="Date">
-    <sv:value>2014-10-21T14:02:27.890+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModifiedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-</sv:node>
