--- magkit-ui/src/main/resources/mgnl-bootstrap/magkit/config.modules.magkit.fieldTypes.sortableMultiValueField.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sv:node sv:name="sortableMultiValueField" xmlns:sv="http://www.jcp.org/jcr/sv/1.0">
-  <sv:property sv:name="jcr:primaryType" sv:type="Name">
-    <sv:value>mgnl:contentNode</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:uuid" sv:type="String">
-    <sv:value>33238fd4-e6af-43cd-89ac-a731404c3d74</sv:value>
-  </sv:property>
-  <sv:property sv:name="definitionClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.SortableMultiValueFieldDefinition</sv:value>
-  </sv:property>
-  <sv:property sv:name="factoryClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.SortableMultiValueFieldFactory</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:createdBy" sv:type="String">
-    <sv:value>admin</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:activationStatus" sv:type="Boolean">
-    <sv:value>false</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:created" sv:type="Date">
-    <sv:value>2015-04-08T09:15:04.122+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:createdBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivated" sv:type="Date">
-    <sv:value>2015-04-08T09:15:04.122+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivatedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModified" sv:type="Date">
-    <sv:value>2015-04-08T09:15:57.067+02:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModifiedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-</sv:node>
