--- magkit-ui/src/main/resources/mgnl-bootstrap/magkit/config.modules.magkit.fieldTypes.specificMultiField.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sv:node sv:name="specificMultiField" xmlns:sv="http://www.jcp.org/jcr/sv/1.0">
-  <sv:property sv:name="jcr:primaryType" sv:type="Name">
-    <sv:value>mgnl:contentNode</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:uuid" sv:type="String">
-    <sv:value>06fdca3b-c056-412d-a6c5-393021bee993</sv:value>
-  </sv:property>
-  <sv:property sv:name="definitionClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.SpecificMultiFieldDefinition</sv:value>
-  </sv:property>
-  <sv:property sv:name="factoryClass" sv:type="String">
-    <sv:value>de.ibmix.magkit.ui.dialogs.fields.SpecificMultiValueFieldFactory</sv:value>
-  </sv:property>
-  <sv:property sv:name="jcr:createdBy" sv:type="String">
-    <sv:value>admin</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:activationStatus" sv:type="Boolean">
-    <sv:value>false</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:created" sv:type="Date">
-    <sv:value>2014-10-28T09:48:40.828+01:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:createdBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivated" sv:type="Date">
-    <sv:value>2014-10-28T09:48:40.829+01:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastActivatedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModified" sv:type="Date">
-    <sv:value>2014-10-28T09:50:29.660+01:00</sv:value>
-  </sv:property>
-  <sv:property sv:name="mgnl:lastModifiedBy" sv:type="String">
-    <sv:value>superuser</sv:value>
-  </sv:property>
-</sv:node>
